Bump oxide.go to latest.
Pull in https://github.com/oxidecomputer/oxide.go/pull/390 to fix an integer
overflow error on large metric values.

+++ receiver/oxidemetricsreceiver/scraper_test.go
@@ -388,7 +388,7 @@ func TestAddHistogram(t *testing.T) {
 									Values: []oxide.Distributionint64{
 										{
 											Bins:   []int{0, 1, 2},
-											Counts: []int{1, 2, 3},
+											Counts: []uint64{1, 2, 3},
 											P50:    1.5,
 											P90:    1.9,
 											P99:    1.99,
@@ -425,7 +425,7 @@ func TestAddHistogram(t *testing.T) {
 									Values: []oxide.Distributiondouble{
 										{
 											Bins:   []float64{0.0, 1.0, 2.0},
-											Counts: []int{1, 2, 3},
+											Counts: []uint64{1, 2, 3},
 											P50:    1.5,
 											P90:    1.9,
 											P99:    1.99,
